requirements.txt
torch
torchvision
grad-cam
opencv-python
matplotlib
pillow
streamlit
numpy
ultralytics
scikit-learn
pandas